requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
attrs==24.2.0
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Django==5.1.4
djangorestframework==3.15.2
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.5
Flask==3.1.0
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
numpy==2.1.3
pandas==2.2.3
pydantic==2.10.1
pydantic_core==2.27.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.21.0
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.41.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
Werkzeug==3.1.3
wheel==0.44.0
